openwrt: Fix config variable name for dtim_period
The vif option dtim_period was accidently renamed dtim_interval in OpenWrt
r38988 ("netifd: add wireless configuration support and port mac80211 to the
new framework"). This is wrong and makes the dtim_period/dtim_interval a dead
option because the rest of the config generation code still uses dtim_period.